+++ src/App.tsx
@@ -1,9 +1,7 @@
 import { Global } from '@emotion/react';
-
 import AppLayout from 'Components/AppLayout';
 import Delete from 'Pages/Delete/Delete';
-import TodoContainer from 'Pages/List/TodoContainer';
-
+import TodoList from 'Pages/List/TodoList';
 import Reset from 'Styles/Reset';
 
 function App() {
@@ -11,7 +9,7 @@ function App() {
     <>
       <Global styles={Reset} />
       <AppLayout>
-        <TodoContainer />
+        <TodoList />
         <Delete />
       </AppLayout>
     </>
